% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flujoclima.R
\docType{data}
\name{flujoclima}
\alias{flujoclima}
\title{Datos de clima y uso vehicular en Autopista Lugones}
\format{
Un data frame con 90 filas y 6 variables:
\describe{
\item{fecha}{fecha, en el formato mes-día}
\item{n}{cantidad de vehículos registrados en el día por el radar RD 171 de la Autopista Lugones, altura ESMA, sentido A}
\item{tavg}{temperatura promedio (°C) del día}
\item{prcp}{nivel de precipitaciones (mm) del día}
\item{dia}{nombre del día}
\item{tipo_dia}{tipo de día (día de semana - fin de semana)}
\item{lluvia}{registro de si la precitación es mayor que 0 o no (llueve - no llueve)}
}
}
\source{
\url{https://github.com/daniellaparada/IC-datasets-docencia/blob/main/fuente/03_test_hipotesis/flujo-clima.csv}
}
\usage{
flujoclima
}
\description{
Datos de clima y de uso vehicular en autopista Lugones, altura ESMA sentido A, en la Ciudad de Buenos Aires, Argentina, desde el 1 de enero de 2023, hasta el 31 de marzo de 2023.
}
\keyword{datasets}
