% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jugadores.R
\docType{data}
\name{jugadores}
\alias{jugadores}
\title{Datos de jugadores de primera división}
\format{
Un data frame con 1105 filas y 10 variables:
\describe{
\item{Equipo}{equipo en el que juega}
\item{Puesto}{posición en la que juega}
\item{Jugador}{nombre del jugador}
\item{Edad}{edad del jugador al 2023}
\item{Nacimiento}{fecha de nacimiento del jugador, en el formato día-mes-año}
\item{Altura}{altura (cm) del jugador}
\item{Ciudad}{ciudad de nacimiento del jugador}
\item{Pais}{país de nacimiento del jugador}
\item{Mes}{mes de nacimiento del jugador}
\item{Anio}{año de nacimiento del jugador}
}
}
\source{
\url{https://daniellaparada.github.io/IC-datasets-docencia/otros-datasets.html}
}
\usage{
jugadores
}
\description{
Datos de jugadores de fútbol de primera división de Argentina al 3 de mayo de 2023.
}
\keyword{datasets}
