% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombresARG.R
\docType{data}
\name{nombresARG}
\alias{nombresARG}
\title{Datos del histórico de nombres en Argentina}
\format{
Un data frame con 9064354 filas y 3 variables:
\describe{
\item{nombre}{nombre registrado, en mayúsculas y sin tilde}
\item{cantidad}{cantidad de nombres registrados en el año}
\item{anio}{año de registro}
}
}
\source{
\url{https://daniellaparada.github.io/IC-datasets-docencia/otros-datasets.html}
}
\usage{
nombresARG
}
\description{
Datos del histórico de nombres registrados ante el Registro Nacional de las Personas en Argentina desde 1922 y hasta 2015 agrupado por nombre y año.
}
\keyword{datasets}
