% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sismos.R
\docType{data}
\name{sismos}
\alias{sismos}
\title{Primavera Dataset}
\format{
A data frame with n rows and m columns.
}
\usage{
data(sismos)
}
\description{
A sample dataset containing some data.
}
\keyword{datasets}
