% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempbici.R
\docType{data}
\name{tempbici}
\alias{tempbici}
\title{Datos de temperatura y uso de EcoBici}
\format{
Un data frame con 333 filas y 7 variables:
\describe{
\item{fecha}{fecha, en el formato año-mes-día}
\item{n}{cantidad de usos de bicletas del día}
\item{tmin}{temperatura mínima (°C) del día}
\item{tmax}{temperatura máxima (°C) del día}
\item{tmed}{temperatura promedio (°C) del día}
\item{dia}{nombre del día}
\item{tipo_dia}{tipo de día (día de semana - fin de semana)}
}
}
\source{
\url{https://github.com/daniellaparada/IC-datasets-docencia/blob/main/fuente/01_regresion/temp-bici.csv}
}
\usage{
tempbici
}
\description{
Datos de temperatura y de uso del sistema EcoBici en la Ciudad de Buenos Aires, Argentina, desde el 1 de enero de 2022, hasta el 31 de diciembre de 2022.
}
\keyword{datasets}
